refactor!: make use of the `mod.rs` files instead of `<module>.rs` files
This helps to organize the code in a better way, and it's more idiomatic
to Rust.

BREAKING CHANGE: this commit changes the way the modules are organized.
