+++ src/object/enviroment.rs
@@ -1,6 +1,6 @@
 use std::{cell::RefCell, collections::HashMap, rc::Rc};
 
-use crate::object::object::Object;
+use crate::object::Object;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct Environment {
